requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
jinja2==3.1.4
pdfplumber==0.11.0
PyMuPDF==1.27.2.2
python-docx==1.1.0
openpyxl==3.1.2
xlrd==2.0.1
striprtf==0.0.26
pytesseract==0.3.13
Pillow==12.2.0
chardet==5.2.0
python-magic==0.4.27
aiofiles==23.2.1
extract-msg==0.52.0
